added CRUD Repository and add is working now

+++ src/main/java/com/thoughtworks/models/BaseEntity.java
@@ -24,14 +24,14 @@ import javax.persistence.MappedSuperclass;
 public class BaseEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  protected Integer id;
+  protected Long id;
 
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
